impl few methods in repo.

+++ spring__backend/FamilyTaskCV/src/main/java/com/konrad/loch/repositories/FamilyRepository.java
@@ -7,11 +7,18 @@ import com.konrad.loch.domains.Family;
 import com.konrad.loch.domains.Father;
 
 public interface FamilyRepository {
-	int createFamily();
+	long createFamily();
+	
 	void addFatherTofamily(Father father, int familyId);
+	
 	void addChildTofamily(Child child, int familyId);
+	
 	List<Integer> searchChild(); //return ids of found childs
+	
 	Family readFamily(int id);
+	
 	Child readChild(int id);
+	
 	Father readFather(int id);
 }
+
